Safari-611.1.10/LayoutTests/platform/ios-13/fast/text/international/system-language/navigator-language/navigator-language-fr-expected.txt

FAIL navigator.language should be fr-fr. Was fr.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
